Comment run.js, move shared code to shared file
